database debugged

+++ src/main/java/hu/progmatic/battleship_torpedotigrisek/repository/UserProfileRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserProfileRepository extends JpaRepository<User, Integer> {
+public interface UserProfileRepository extends JpaRepository<User, Long> {
     User findByName(String userName);
 }
